Enabling to run with auto-gptq

+++ src/transformers/models/llama/modeling_llama.py
@@ -332,7 +332,7 @@ class LlamaAttention(nn.Module):
         key_states = repeat_kv(key_states, self.num_key_value_groups)
         value_states = repeat_kv(value_states, self.num_key_value_groups)
 
-        attn_weights = torch.matmul(query_states, key_states.transpose(2, 3)) / math.sqrt(self.head_dim)
+        attn_weights = torch.matmul(query_states.float(), key_states.transpose(2, 3).float()) / math.sqrt(self.head_dim)
 
         #if attn_weights.size() != (bsz, self.num_heads, q_len, kv_seq_len):
         #    raise ValueError(
@@ -349,7 +349,7 @@ class LlamaAttention(nn.Module):
 
         # upcast attention to fp32
         attn_weights = nn.functional.softmax(attn_weights, dim=-1, dtype=torch.float32).to(query_states.dtype)
-        attn_output = torch.matmul(attn_weights, value_states)
+        attn_output = torch.matmul(attn_weights.float(), value_states.float())
 
         #if attn_output.size() != (bsz, self.num_heads, q_len, self.head_dim):
         #    raise ValueError(
